Test: Split perf core benchmarks
- convert to pytest
- split into graph, mifs, pick
